+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - Added `MigrationTrait::dropColumnIfExists()` and `MigrationTrait:dropIndexIfExists()`
 - Enhanced `NavBar::$languageRoute` to be merged with query parameters
 - Enhanced `UrlManager` to accept a `defaultLanguage` for each URL created
-- Fixed `UrlManager` to correctly redirect default language URLs while keeping the query parameters
+- Fixed `UrlManager::setApplicationLanguage()` to keep the query parameters when redirecting default language URLs
+- Fixed `UrlManager::getImmutableRuleParams()` to include allowed URL characters
 
 ## 2.3.1 (Apr 15, 2024)